+++ src/types/session.ts
@@ -1,7 +1,8 @@
 /**
  * Platform the game is currently being played on.
  */
-export type Platform = 'wortal' | 'link' | 'viber' | 'gd' | 'facebook' | 'crazygames' | 'gamepix' | 'telegram' | 'debug'
+export type Platform = 'wortal' | 'link' | 'viber' | 'gd' | 'facebook' | 'crazygames' | 'gamepix' | 'telegram' |
+    'gamemonetize' | 'debug'
 
 /**
  * Device the player is currently using.